partsdb/backend/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
django-filter==24.3
django-cors-headers==4.3.1
python-dotenv==1.0.1
pydantic==2.9.2
requests==2.32.3
drf-spectacular==0.27.2
celery==5.4.0
redis==5.0.8
psycopg[binary]==3.2.10
pytest==8.3.2
pytest-django==4.8.0
factory-boy==3.3.0
PyYAML==6.0.1